+++ freeway-benchmark/src/main/java/com/jujin/freeway/benchmarks/client/WsClient.java
@@ -1,6 +1,9 @@
 package com.jujin.freeway.benchmarks.client;
 
-import java.io.*;
+import java.io.BufferedInputStream;
+import java.io.DataInputStream;
+import java.io.IOException;
+import java.io.OutputStream;
 import java.net.Socket;
 import java.nio.charset.StandardCharsets;
 import java.security.SecureRandom;